First pass of Bank 84 rewrite and documentation, Bank 85 message boxes cleaned up
- Documented and rewrote Bank 84
  - This is a first pass, and needs a second read through
- Cleaned up and made the message boxes in Bank 85 more concise
- get_DoorDef() door_direction_ field fixed
- HUD_RAM_OFFSET macro
- Block Types and BTS Types enumerations
- Lava Stage struct
- Various other improvements and documentations

+++ src/sm_93.c
@@ -150,7 +150,7 @@ uint16 GetProjectileTrailFrame(uint16 proj_index) {  // 0x9381D1
   if (projectile_instruction_timers[proj_index >> 1] == 1 && !IS_FUNC(get_ProjectileInstr(instr_ptr).func_ptr))
       instr_offset = 0;
   else /* (projectile_instruction_timers[proj_index >> 1] != 1 || IS_FUNC(get_ProjectileInstr(instr_ptr).func_ptr)) */
-      instr_offset = -sizeof(ProjectileInstr);
+      instr_offset = -(int)sizeof(ProjectileInstr);
   return get_ProjectileInstr(instr_ptr + instr_offset).trail_frame;
 }
